tests: add diffing logic to tests
Since the resulting string from small-measurements.txt is hundreds of
kBs large, dumping expected and actual to stdout does nothing to help
find the what's different. Added parsing logic which:

1. Validates that the output is in the correct shape (the validation is
   kind of scuffed, just a bunch of `.unwrap()`s scattered all over)
2. Validates that the resulting maps are identical, for each discrepancy
   (value missing, value different) prints them to stdout.

This already made one bug very clear: The parallel code is sometimes
getting a different mean (mostly off by .1), probably due to the
difference in order of operations

NOTE: the diffing logic is kind of crude, and doesn't cover all cases
(e.g. when actual_map has a key which is not present in expected_map),
since it's only intended to speed up the process of finding the
difference. If I find it necessary we can add more functionality later.