change counters from float to integers
There was a bug where plain and performant sometimes disagreed on values
due to floating point rounding and so on. In order to fix this, values
are all treated as integers and only converted to floats when printing,
this ensures that `record.acc` is *always* the correct value with no
rounding errors.

NOTE: had to update data/small-ref.txt because in the previous version
of the code, 0.0 would sometimes be formatted as `-0.0`, which makes no
sense. The new version does not have this bug.