refactor: removed globals/thread locals
Due to the nature of globals and thread locals, unrelated tests were
failing, because one test would influence the other through thread
locals and globals. In order to fix this, all state is now stored inside
structs which are passed around. ExecutionCtx is the context which is
common to both joiners and readers, and is passed around as a pointer
because:
1. Using any of the safe wrappers Arc, Rc, etc. would have performance
   implications
2. It's always valid to use from any thread because it's only freed at
   the end of entrypoint, when all threads are guaranteed to have
   finished already.