optimization: reduce buf size
This changed the time from 15.5s to 12.3s, pretty good! Even though
`buf.len()` is now not a power of 2, this doesn't matter in practice
because we were reading a *lot* of unnecessary bytes. Each line is at
most 100 and something characters, so we have enough space for the
worst-case scenario, which is that the line's first character is exactly
at buf[chunk_size - 1].